[BUGFIX] DataCapture Component Tick Logic

+++ GGI_Project/Source/GGI/Public/HandMotionCaptureComponent.h
@@ -39,7 +39,7 @@ private:
 
 	int32 LSTMTimeStep;  // 40
 
-	int32 AccumulatedDeltaTime;
+	float AccumulatedDeltaTime;
 
 private:	
 	UPROPERTY()
